chore: move sanitization util to a new file
